pyproject.toml
[project]
name = "spear-rag"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.4",
    "fastapi>=0.121.1",
    "langchain>=1.0.5",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.0",
    "langchain-openai>=1.0.2",
    "openai>=2.7.2",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "streamlit-lottie>=0.0.5",
    "uvicorn[standard]>=0.38.0",
]
